requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.0.1
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
WTForms==3.0.1
